+++ packages/react-web-kit/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.2](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@1.12.1...@tecsinapse/react-web-kit@1.12.2) (2021-12-22)
+
+**Note:** Version bump only for package @tecsinapse/react-web-kit
+
+
+
+
+
+## [1.12.1](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@1.12.0...@tecsinapse/react-web-kit@1.12.1) (2021-12-21)
+
+**Note:** Version bump only for package @tecsinapse/react-web-kit
+
+
+
+
+
+# [1.12.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@1.11.1...@tecsinapse/react-web-kit@1.12.0) (2021-12-21)
+
+**Note:** Version bump only for package @tecsinapse/react-web-kit
+
+
+
+
+
+## [1.11.1](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@1.11.0...@tecsinapse/react-web-kit@1.11.1) (2021-12-20)
+
+**Note:** Version bump only for package @tecsinapse/react-web-kit
+
+
+
+
+
 # [1.11.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@1.10.5...@tecsinapse/react-web-kit@1.11.0) (2021-12-17)
 
 **Note:** Version bump only for package @tecsinapse/react-web-kit